backend/pyproject.toml
[project]
name = "fastapi-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.7",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
]
